requirements.txt
Django
django-cors-headers
djangochannelsrestframework
djangorestframework
psycopg2-binary
sqlparse
channels
channels-redis
redis
